# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropTech Platform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rental/tenants:
    get:
      operationId: listTenants
      tags: [rental]
      summary: List tenants
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [rental]
      summary: Create tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTenantBody"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /rental/tenants/{id}:
    get:
      operationId: getTenant
      tags: [rental]
      summary: Get tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateTenant
      tags: [rental]
      summary: Update tenant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTenantBody"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # RENTAL - LEASE CONTRACTS
  /rental/contracts:
    get:
      operationId: listLeaseContracts
      tags: [rental]
      summary: List lease contracts
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lease contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaseContract"
    post:
      operationId: createLeaseContract
      tags: [rental]
      summary: Create lease contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaseContractBody"
      responses:
        "201":
          description: Lease contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaseContract"
  /rental/contracts/{id}:
    get:
      operationId: getLeaseContract
      tags: [rental]
      summary: Get lease contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lease contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaseContract"
    patch:
      operationId: updateLeaseContract
      tags: [rental]
      summary: Update lease contract
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaseContractBody"
      responses:
        "200":
          description: Updated lease contract
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaseContract"

  # RENTAL - ACCRUALS
  /rental/accruals:
    get:
      operationId: listAccruals
      tags: [rental]
      summary: List accruals
      parameters:
        - name: leaseContractId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of accruals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accrual"
  /rental/accruals/recalculate:
    post:
      operationId: recalculateAccruals
      tags: [rental]
      summary: Recalculate accruals for a lease contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecalculateAccrualsBody"
      responses:
        "200":
          description: Recalculated accruals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Accrual"

  # RENTAL - PAYMENTS
  /rental/payments:
    get:
      operationId: listPayments
      tags: [rental]
      summary: List rental payments
      parameters:
        - name: leaseContractId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
    post:
      operationId: createPayment
      tags: [rental]
      summary: Record rental payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # RENTAL - DEPOSITS
  /rental/deposits:
    get:
      operationId: listDeposits
      tags: [rental]
      summary: List deposits
      parameters:
        - name: leaseContractId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of deposits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deposit"
    post:
      operationId: createDeposit
      tags: [rental]
      summary: Record deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepositBody"
      responses:
        "201":
          description: Deposit recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /rental/deposits/{id}:
    patch:
      operationId: updateDeposit
      tags: [rental]
      summary: Update deposit status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepositBody"
      responses:
        "200":
          description: Updated deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # RENTAL - EXPENSES
  /rental/expenses:
    get:
      operationId: listExpenses
      tags: [rental]
      summary: List rental expenses
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of expenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [rental]
      summary: Record expense
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseBody"
      responses:
        "201":
          description: Expense recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"

  # RENTAL - PROPERTIES (with rental status)
  /rental/properties:
    get:
      operationId: listRentalProperties
      tags: [rental]
      summary: List properties with rental info
      parameters:
        - name: rentalStatus
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of rental properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalProperty"
  /rental/properties/{id}/activate:
    post:
      operationId: activateRentalProperty
      tags: [rental]
      summary: Activate property for rental
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property activated for rental
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalProperty"
  /rental/properties/{id}/performance:
    get:
      operationId: getPropertyPerformance
      tags: [rental]
      summary: Get property rental performance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property performance metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyPerformance"

  # RENTAL - STATEMENTS
  /rental/statements:
    get:
      operationId: listOwnerStatements
      tags: [rental]
      summary: List owner statements
      parameters:
        - name: propertyId
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of owner statements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnerStatement"
  /rental/statements/generate:
    post:
      operationId: generateOwnerStatement
      tags: [rental]
      summary: Generate owner statement for month
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateStatementBody"
      responses:
        "200":
          description: Generated owner statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OwnerStatement"

  # DASHBOARD
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Tenant:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        iin:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - fullName
        - status
        - createdAt
        - updatedAt
    CreateTenantBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        iin:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        comment:
          type: ["string", "null"]
      required:
        - fullName
        - status
    UpdateTenantBody:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        iin:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, inactive, blacklisted]
        comment:
          type: ["string", "null"]
    LeaseContract:
      type: object
      properties:
        id:
          type: integer
        propertyId:
          type: integer
        propertyUnitNumber:
          type: ["string", "null"]
        tenantId:
          type: integer
        tenantName:
          type: ["string", "null"]
        contractNumber:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        rentAmount:
          type: number
        currency:
          type: string
        depositAmount:
          type: ["number", "null"]
        accrualDay:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated]
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - propertyId
        - tenantId
        - contractNumber
        - startDate
        - rentAmount
        - currency
        - status
        - createdAt
        - updatedAt
    CreateLeaseContractBody:
      type: object
      properties:
        propertyId:
          type: integer
        tenantId:
          type: integer
        contractNumber:
          type: string
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        rentAmount:
          type: number
        currency:
          type: string
        depositAmount:
          type: ["number", "null"]
        accrualDay:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated]
        comment:
          type: ["string", "null"]
      required:
        - propertyId
        - tenantId
        - contractNumber
        - startDate
        - rentAmount
        - currency
        - status
    UpdateLeaseContractBody:
      type: object
      properties:
        startDate:
          type: string
        endDate:
          type: ["string", "null"]
        rentAmount:
          type: number
        currency:
          type: string
        depositAmount:
          type: ["number", "null"]
        accrualDay:
          type: ["integer", "null"]
        status:
          type: string
          enum: [draft, active, expired, terminated]
        comment:
          type: ["string", "null"]
...
    RecalculateAccrualsBody:
      type: object
      properties:
        leaseContractId:
          type: integer
        fromDate:
          type: ["string", "null"]
      required:
        - leaseContractId
...
    CreateDepositBody:
      type: object
      properties:
        leaseContractId:
          type: integer
        amount:
          type: number
        currency:
          type: string
        receivedDate:
          type: string
        note:
          type: ["string", "null"]
      required:
        - leaseContractId
        - amount
        - currency
        - receivedDate
...